backend/requirements.txt
django
dotenv
djangorestframework
django-allauth
dj-rest-auth
